Fixes issues with compiling (Piazza @327)
* See [Piazza note](https://piazza.com/class/kmwft7y9xzm6cb?cid=327)

+++ FileService.cpp
@@ -29,8 +29,8 @@ FileService::~FileService() {
 }
 
 bool FileService::endswith(string str, string suffix) {
-  int pos = str.rfind(suffix);
-  return pos == (str.length() - suffix.length());
+  size_t pos = str.rfind(suffix);
+  return pos == str.length() - suffix.length() ;
 }
 
 void FileService::get(HTTPRequest *request, HTTPResponse *response) {
